Added UserValidator //

+++ Entities/Concrete/Brand.cs
@@ -9,7 +9,7 @@ namespace Entities.Concrete
 {
     public class Brand:IEntity
     {
-        public int Id { get; set; }
+        public Byte Id { get; set; }
         public string BrandName { get; set; }
     }
 }
