+++ Entities/Concrete/Car.cs
@@ -9,9 +9,9 @@ namespace Entities.Concrete
 {
     public class Car:IEntity
     {
-        public int Id { get; set; }
-        public int BrandId { get; set; }
-        public int ColorId { get; set; }
+        public short Id { get; set; }
+        public Byte BrandId { get; set; }
+        public Byte ColorId { get; set; }
         public string CarName { get; set; }
         public DateTime ModelYear { get; set; }
         public decimal DailyPrice { get; set; }
