+++ Entities/Concrete/Customer.cs
@@ -9,8 +9,8 @@ namespace Entities.Concrete
 {
     public class Customer : IEntity
     {
-        public int Id { get; set; }
-        public int UserId { get; set; }
+        public short Id { get; set; }
+        public short UserId { get; set; }
         public string CompanyName { get; set; }
     }
 }
