+++ Entities/Concrete/Rental.cs
@@ -9,9 +9,9 @@ namespace Entities.Concrete
 {
     public class Rental:IEntity
     {
-        public int Id { get; set; }
-        public int CarId { get; set; }
-        public int CustomerId { get; set; }
+        public short Id { get; set; }
+        public short CarId { get; set; }
+        public short CustomerId { get; set; }
         public DateTime RentDate { get; set; }
         public DateTime ReturnDate { get; set; }
     }
